Add files via upload
automating image labeling through the process of copying the files with a jpeg extension with the same prefix as the jpeg. Test.py opens the image corresponding XML file to use color detection to determine the bounding box coordinates within a user-defined area and allow user to set label class written into the XML file.